move path and node utility methods into graph module
